requirements.txt
streamlit
qrcode
Pillow
# pycairo
numpy
cairocffi